requirements.txt
streamlit
speechrecognition
pyttsx3
opencv-python
requests
pillow
pygame